WP and Android game made for the NASA App Challenge
Space Explorers is a game that allows users to compete between each
other to gain the score lead by showing their knowledge about their
local zone, their country, the earth and even the universe. The app was
design for Windows Phone and Android devices. The main objective is to
teach the people about the information NASA has of their local
comunities and the universe

You will start the game with local questions about geography and data
recolected by NASA and given by the user through the "Question Factory".
Here the user will use his knowledge about the culture and geography
that surrounds them. The objective of this part is that the user learns
how to use the game as well as giving him information he probably didn't
knew about his area. Later on, the area of the questions will grow into
country size and so on until the questions about the universe appear. We
created a map in which the user has to answer questions in order to pass
to the next level. Each of the questions has a particular category that
is represented by the color of the stage.